+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - PR #3070 Move functions.h and related source to legacy
 - PR #2951 Allow set_index to handle a list of column names
 - PR #3067 Java method to return size of device memory buffer
+- PR #3135 Add nvtx utilities to cudf::nvtx namespace
 
 ## Bug Fixes